Regenerate regression tests artefacts (Marriage abroad) This commit shows the affected/changes to regression test artefacts as a result of removing the condition added previously to redirect Kyrgyzstan to Kazakhstan consular fees link.
I have reviewed the changes to the regression test artefacts and it is as expected.

The removal affects only opposite sex (marriages in the uk and in ceremony country) outcomes in Kazakhstan and Kyrgyzstan.

The affect link does not appear on the third country routes and hence is not affected and as expected.

Already existing integration tests passed, giving assurance that the aforementioned changes have had no affect to the flow.