Content amendments to Check UK Visa
Here I have made changes as required in this document from the Content team.

These changes became necessary as the previous changes from commits 558c3af...de400ee,

did not meet the expectations of the home office hence the need to make further changes here in this commit

and the ones to following this.

Essentially, the idea here is to further split and make explicit the composition and definition of Common Travel Areas.

This only contains Content amends and to meet the aforementioned purpose.

You will observe that I have renamed outcome_transit_to_the_cta.govspeak.erb to

outcome_transit_to_the_republic_of_ireland.govspeak.erb, this simply is following the Content amends to reflect it meaning with the file's name.

I have updated the checksum data to properly reflect the changed to the questions and outcome files.

Consequently, I have also run the tests (bundle exec rake) and 32 tests, 28 of those all related to the transit route, 1 for checksum and 1 for regressions artefacts.

This I believe (at this point) is as expected.